Minimum Functionality
This is the source code, make file and /bin (executable locale), and sample output (/images) for the baseline functionality for this project.  The goal is to have a user trigger data collection from a FLIR BFS camera.  Data shall be saved as a series of PPM images, currently in the "images" directory.  Running Command_PPM_Capture will lead to the user being prompted to press SPACE to start logging images.  Images shall be logged for 60 seconds; this is currently hard coded, but should become parameterized.  Only every 5th image is saved as the save process was taking too long for the duration of a "mission" to be logged start to end.  The tail end of the images were not being processed within the 60 (or however many) seconds of a run.