ROS publisher and subscriber base code
These were created within a workspace called, "catkin_ws" within the Home directory of ROS316.  They are source files residing in /catkin_ws/src/beginner_tutorials/src.  Also in this commit are the CMakeLists.txt, and package.xml files (stored in the /beginner_tutorials package directory), which are used for building these nodes in this package.

bool_pub publishes a String, not a Bool at this point.  However, the strings published are "0" and "1" to fake boolean values until such time as the type conversion is made.  The pulisher starts by publishing "0" (topic "record"), which is interpreted by the subscriber to mean do not record camera data.  The user may then press SPACE to begin publishing "1", which should mean, start recording camera data.  Pressing "q" reverts to not recording (quit recording).

bool_sub listens to the record topic, and prints what it hears.  It does not yet interface with FLIR API code.  That is the next step of development.