Prototype code for acquiring specific camera
Loop until a camera with a specified serial number is acquired.  Implemented in bool_sub, this will allow a specific launched node to be attached to a specific camera, as identified by serial number.  The serial number will ultimately be specified by setting a ROS param in a launch file.  A bool_sub node will be launched with a specific SN associated with it, and will thus be responsible for acquiring and saving images from only the camera with that SN.