bool_sub for SN specific node
A version of bool_sub that works with a specific camera (if it is found) with a specific serial number (SN).  Only images for that camera are captured.  Other cameras will be handled by other nodes (copies of bool_sub that are launched).  Currently the desired SN is hard coded in, but it should ultimately be handled by a ROS param.  With two FLIR cameras plugged in simultaneously, the system currently identifies the second camera as being the desired device, even though the SN is different.  However, this is being developed for a FLIR POE, and a Xenic Gobi, so the problem may ultimately not exist for that pairing.  Further testing required.