Working Copy
Tested successfully for one camera, the following uses the camera_sub launch file to launch bool.sub with the parameter "desiredSN" to set the serial number of the camera associated with the node being launched to some specified value.  bool_sub then looks for this value in the camera list that it finds, and if it finds it, starts recording images for that camera if and only if it receives a value of "1" from the "record" topic published by bool_pub.  Images are saved to /home/Data/ and contain the serial number, and date and time in the file name.  They are PPM images.